ql_step1
rules: yellow box has to reach one of the goals to end the game (green or red cell)
rewards: each step gives a negative reward of -0.04. the red cell gives a negative reward of -1. the green one gives a positive reward of +1
states: each cell is a state the agent can be
actions: there are only 4 actions (up down right left)